fix bugs through unit tests.
